airflow/requirements.txt
apache-airflow==2.10.5
kafka-python==2.0.6